New `types/numeric.ts`; applied to `handlers/numeric.ts`
