+++ server/shared/types/index.ts
@@ -5,3 +5,5 @@ export * from './termdb'
 export * from './filter'
 export * from './vocab'
 export * from './categorical'
+export * from './condition'
+export * from './numeric'
