+++ server/shared/types/termsetting.ts
@@ -125,42 +125,42 @@ export interface TermSettingOpts extends BaseTermSettingOpts {
 }
 
 /*** types and interfaces supporting TermSettingInstance type ***/
-type NumObjRangeEntry = any //{}
+// type NumObjRangeEntry = any //{}
 
-export type BrushEntry = {
-	//No documentation!
-	orig: string
-	range: {
-		start: number
-		stop: number
-	}
-	init: () => void
-}
+// export type BrushEntry = {
+// 	//No documentation!
+// 	orig: string
+// 	range: {
+// 		start: number
+// 		stop: number
+// 	}
+// 	init: () => void
+// }
 
-type DensityData = {
-	maxvalue: number
-	minvalue: number
-}
+// type DensityData = {
+// 	maxvalue: number
+// 	minvalue: number
+// }
 
-type PlotSize = {
-	width: number
-	height: number
-	xpad: number
-	ypad: number
-}
+// type PlotSize = {
+// 	width: number
+// 	height: number
+// 	xpad: number
+// 	ypad: number
+// }
 
-export type NumberObj = {
-	binsize_g?: any //dom element??
-	brushes?: BrushEntry[]
-	density_data?: DensityData
-	no_density_data?: true
-	plot_size?: PlotSize
-	ranges?: NumObjRangeEntry[]
-	svg?: any
-	xscale?: any
-}
+// export type NumberObj = {
+// 	binsize_g?: any //dom element??
+// 	brushes: BrushEntry[]
+// 	density_data: DensityData
+// 	no_density_data?: true
+// 	plot_size: PlotSize
+// 	ranges?: NumObjRangeEntry[]
+// 	svg?: any
+// 	xscale?: any
+// }
 
-type InstanceDom = {
+export type InstanceDom = {
 	//Separate from the Dom outlined in termsetting.ts?????
 	//Required
 	holder: any
@@ -195,7 +195,6 @@ export type TermSettingInstance = {
 	handlerByType?: { [index: string]: Handler }
 	hasError?: boolean
 	noTermPromptOptions?: NoTermPromptOptsEntry[]
-	num_obj?: NumberObj
 	numqByTermIdModeType?: any
 	opts: TermSettingOpts
 	placeholder: string | undefined
